+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.7 - 2019-12-06
+
+### Changed
+
+-   note model return user even if disabled
+
+### Added
+
+-   notes service getNotes now accepts an array of conditions to filter notes on.
+
 ## 1.0.6 - 2019-08-29
 
 ### Changed